+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Option to configure trusted_proxies for which the X-Forwarded-For and X-Real-IP headers are used as client IP (b062edca)
 - Added multiple unit tests for ensuring correct functionality
+- Add `excluded_logs` config option to skip specific logs and operators. Works both for classic and tiled logs.
 
 ### Changed
 - Updated weak cipher suites to stronger ones (939517cd)